Revert "Merge remote-tracking branch 'stevenlovegrove/master'"
This reverts commit 245439e242450e29ad22f2e08083aa7cded6af6b, reversing
changes made to 2984fc7b7b631677fa3aecf2d2c5d75f6bab04a1.

+++ include/pangolin/compat/glutbitmap.h
@@ -28,10 +28,10 @@
 #ifndef PANGOLIN_COMPAT_GLUT_BITMAP_H
 #define PANGOLIN_COMPAT_GLUT_BITMAP_H
 
-#include <pangolin/gl/glglut.h>
+#include <pangolin/glglut.h>
 
 #ifndef HAVE_GLUT
-#include <pangolin/gl/glfont.h>
+#include <pangolin/glfont.h>
 
 #ifdef HAVE_GLES
 GLfloat g_raster_pos[4];
@@ -89,7 +89,7 @@ inline void glutBitmapString(void *font, const unsigned char *str)
 
 inline int glutBitmapLength(void *font, const unsigned char *str)
 {
-    return (int)(pangolin::GlFont::I().Text((const char *)str).Width());
+    return pangolin::GlFont::I().Text((const char *)str).Width();
 }
 
 #define GLUT_BITMAP_HELVETICA_12 0;
